Move components for setup to setup dir
