+++ src/index.js
@@ -2,8 +2,4 @@
 // eslint-disable-next-line import/export
 export * as amphtml from './amphtml';
 export * as helpers from './helpers';
-
-export AmpScript from './AmpScript';
-export AmpScripts from './AmpScripts';
-export AmpScriptsManager from './AmpScriptsManager';
-export headerBoilerplate from './headerBoilerplate';
+export * as setup from './setup';
