+++ src/setup/AmpScriptsManager.js
