+++ src/setup/index.js
@@ -1,0 +1,4 @@
+export AmpScript from './AmpScript';
+export AmpScripts from './AmpScripts';
+export AmpScriptsManager from './AmpScriptsManager';
+export headerBoilerplate from './headerBoilerplate';
